+++ js/util.js
@@ -1,0 +1,20 @@
+// модуль с вспомогательными функциями
+
+// Функция для проверки максимальной длины строки.
+const checkStringLength = (string, length) => (string.length <= length);
+
+// Функция, возвращающая случайное целое число из переданного диапазона включительно
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+// возвращает случайный элемент массива
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length -1)];
+
+export {
+  checkStringLength,
+  getRandomPositiveInteger,
+  getRandomArrayElement};
